+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### New features
 
-* `Add [`clj-reload`](https://github.com/tonsky/clj-reload) functionality by implementing new `cider.clj-reload/reload`, `cider.clj-reload/reload-all`, `cider.clj-reload/reload-clear` ops.
+* Add [`clj-reload`](https://github.com/tonsky/clj-reload) functionality by implementing new `cider.clj-reload/reload`, `cider.clj-reload/reload-all`, `cider.clj-reload/reload-clear` ops.
   * These are drop-in replacements for `refresh`, `refresh-all`, `refresh-clear`.
 
 ### Changes